refactor(main): moved file widgets out of main into views
